# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "MiniSense RESTful API Documentation",
        "contact": {
            "email": "alefs07@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3333/",
            "description": "MiniSense RESTful API test"
        }
    ],
    "paths": {
...
        "/sensorDevice": {
            "post": {
                "summary": "Registration of a sensory device",
                "description": "Route responsible for registering a sensory device in a user's account through their email. No parameters are needed in the request but you must assign a body in json format with: email, label and description. In response, if the body email does not exist in the database or if the label already exists in the user's device list you will receive a 400 response, but if the email exists in the database and the label does not exist you will receive a 200 response with a json with: id, key, label and description.",
                "tags": ["Sensor Device"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SensorDevice"
                            },
                            "examples": {
                                "sensorDevice": {
                                    "value": {
                                        "email": "my_email@gmail.com",
                                        "label": "my label",
                                        "description": "my description"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "User not found or sensor device already exists."
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/SensorDevice"
                                },
                                "examples": {
                                    "sensorDevice": {
                                        "value": {
                                            "id": 1,
                                            "key": "df43d6ab-47a0-46b5-a403-ba46297c1558",
                                            "label": "my label",
                                            "description": "my description"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "List of all sensory devices of registered users",
                "description": "Route responsible for listing all available sensory devices in the database. No parameters are needed in the request and no body. As a response you will receive a 200 response with a json with an array with all the sensory devices registered or an empty array.",
                "tags": ["Sensor Device"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "$ref": "#/components/schemas/SensorDevice"
                                },
                                "examples": {
                                    "sensorDevice": {
                                        "value": [
                                            {
                                                "id": 1,
                                                "key": "88705450-f3b5-480a-accc-d8ee07bb232f",
                                                "label": "my label 01",
                                                "description": "my description 01",
                                                "user_id": 1
                                            },
                                            {
                                                "id": 2,
                                                "key": "70a5c190-352e-4a96-b2d6-feccaefff5ff",
                                                "label": "my label 02",
                                                "description": "my description 02",
                                                "user_id": 4
                                            },
                                            {
                                                "id": 3,
                                                "key": "2d6969ed-0892-4759-b9c5-7cbf03b80ddc",
                                                "label": "my label 03",
                                                "description": "my description 03",
                                                "user_id": 9
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sensorDevice/device/{key}": {
            "get": {
                "summary": "Query a sensor device by its key",
                "description": "Route responsible for querying a sensory device by its key. You must pass the key of a sensory device as a parameter. A bodysuit is not required. As an answer, if the key is not compatible with another key in the database, you will receive a 400 response, but if the key is compatible you will receive a 200 response with a json with the device data: id, key, label, description and streams array. The streams array contains: id, key, label, unitId, deviceId, measurementsCount and measurements array. Only the last 5 measurements will be displayed. To view all measurements, a data stream must be consulted by its key.",
                "tags": ["Sensor Device"],
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Device key for query",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Sensor device not found."
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/SensorDevice"
                                },
                                "examples": {
                                    "sensorDevice": {
                                        "value": {
                                            "id": 1,
                                            "key": "8e9a5e56-cbe0-4e52-8177-4896a95ad3ac",
                                            "label": "my device label 01",
                                            "description": "my description 01",
                                            "streams": {
                                                "id": 1,
                                                "key": "19252280-6df3-4397-8138-38e74584ef1b",
                                                "label": "my stream label 01",
                                                "unitId": 1,
                                                "deviceId": 1,
                                                "measurementCount": 2,
                                                "measurements": [
                                                    {
                                                        "timestamp": "2022-20-01 20:33:15",
                                                        "value": 259.43
                                                    },
                                                    {
                                                        "timestamp": "2022-20-01 20:33:30",
                                                        "value": 260.12
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sensorDevice/userDevice/{email}": {
            "get": {
                "summary": "Query a user's sensory devices via their email",
                "description": "Route responsible for consulting a user's sensory devices via their email. You must pass a user's email as a parameter. A bodysuit is not required. As a response, if the email is not compatible with another email in the database, you will receive a 400 response, but if the email is compatible you will receive a 200 response with a json with an array with all the data of the user's devices: id, key, label, description and streams array. The streams array contains: id, key, label, unitId, deviceId and measurementsCount.",
                "tags": ["Sensor Device"],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User email for device consultation",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "User not found."
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/SensorDevice"
                                },
                                "examples": {
                                    "sensorDevice": {
                                        "value": [
                                            {
                                                "id": 1,
                                                "key": "70444e39-8d29-49e8-9513-2b165eec4b28",
                                                "label": "my device label 01",
                                                "description": "my description 01",
                                                "streams": [
                                                    {
                                                        "id": 1,
                                                        "key": "c40a5b48-ae5d-4c76-8dd2-3a095201c628",
                                                        "label": "my stream label 01",
                                                        "unitId": 4,
                                                        "deviceId": 1,
                                                        "measurementCount": 12
                                                    }
                                                ]
                                            },
                                            {
                                                "id": 2,
                                                "key": "67445649-20f5-42a2-80f0-f7186e3f28a5",
                                                "label": "my device label 02",
                                                "description": "my description 02",
                                                "streams": [
                                                    {
                                                        "id": 1,
                                                        "key": "ff05fe09-3efd-4b5f-acda-db073e13a482",
                                                        "label": "my stream label 02",
                                                        "unitId": 7,
                                                        "deviceId": 2,
                                                        "measurementCount": 3
                                                    }
                                                ]
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SensorDevice": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "key": {
                        "type": "string"
                    },
                    "label": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "number"
                    }
                }
                
            },
...
        }
    }
}